Add parallel diagnosis helpers and module processing support
- Introduced `parallel_diagnose_helpers.go` to categorize files based on import dependencies and track performance metrics for parallel processing.
- Added `parallel_diagnose_module.go` to enhance module result enrichment, allowing for shared scope diagnostics across multi-file modules.
- Updated `parallel_diagnose.go` to integrate new helper functions and streamline the diagnosis process for directories with multi-file modules.
- Improved type checking and diagnostics for module-related operations, enhancing overall performance and accuracy in the language's processing capabilities.

This update significantly enriches the language's capabilities for parallel diagnosis and module handling, improving efficiency and accuracy in diagnostics.
