+++ internal/driver/parallel_diagnose.go
@@ -2,12 +2,9 @@ package driver
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"path/filepath"
 	"runtime"
 	"sort"
-	"sync/atomic"
 
 	"golang.org/x/sync/errgroup"
 
@@ -19,139 +16,7 @@ import (
 	"surge/internal/sema"
 	"surge/internal/source"
 	"surge/internal/symbols"
-	"surge/internal/types"
 )
-
-// FileClass categorizes files based on their import dependencies
-type FileClass int
-
-const (
-	FileDependent        FileClass = iota // Imports project modules
-	FileStdlibOnly                        // Only imports stdlib modules
-	FileFullyIndependent                  // No imports
-)
-
-// parallelMetrics tracks performance metrics for parallel processing
-type parallelMetrics struct {
-	// Worker pool metrics
-	workersActive    atomic.Int32 // Currently running workers
-	workersCompleted atomic.Int64 // Total completed tasks
-	workersErrors    atomic.Int64 // Total errors encountered
-
-	// Cache metrics
-	cacheHits   atomic.Int64 // Memory cache hits
-	cacheMisses atomic.Int64 // Memory cache misses
-	diskHits    atomic.Int64 // Disk cache hits
-	diskMisses  atomic.Int64 // Disk cache misses
-
-	// File classification metrics
-	filesIndependent atomic.Int64 // Files with no imports
-	filesStdlibOnly  atomic.Int64 // Files only importing stdlib
-	filesDependent   atomic.Int64 // Files with project dependencies
-
-	// Batch parallelism metrics
-	batchCount     atomic.Int64 // Number of batches processed
-	batchSizeTotal atomic.Int64 // Total module count across all batches
-	batchSizeMax   atomic.Int64 // Largest batch size
-}
-
-// emitMetrics outputs all collected metrics at the end of processing
-func (pm *parallelMetrics) emitMetrics() string {
-	// Worker stats
-	completed := pm.workersCompleted.Load()
-	errs := pm.workersErrors.Load()
-
-	// Cache stats
-	memHits := pm.cacheHits.Load()
-	memMisses := pm.cacheMisses.Load()
-	memTotal := memHits + memMisses
-	memHitRate := 0.0
-	if memTotal > 0 {
-		memHitRate = float64(memHits) / float64(memTotal) * 100
-	}
-
-	diskHits := pm.diskHits.Load()
-	diskMisses := pm.diskMisses.Load()
-	diskTotal := diskHits + diskMisses
-	diskHitRate := 0.0
-	if diskTotal > 0 {
-		diskHitRate = float64(diskHits) / float64(diskTotal) * 100
-	}
-
-	// File classification stats
-	independent := pm.filesIndependent.Load()
-	stdlibOnly := pm.filesStdlibOnly.Load()
-	dependent := pm.filesDependent.Load()
-	totalFiles := independent + stdlibOnly + dependent
-
-	// Batch parallelism stats
-	batchCount := pm.batchCount.Load()
-	batchTotal := pm.batchSizeTotal.Load()
-	batchMax := pm.batchSizeMax.Load()
-	batchAvg := 0.0
-	if batchCount > 0 {
-		batchAvg = float64(batchTotal) / float64(batchCount)
-	}
-
-	return fmt.Sprintf(
-		"workers: %d completed, %d errors | "+
-			"cache: mem=%d/%d (%.1f%%), disk=%d/%d (%.1f%%) | "+
-			"files: %d total (%d indep, %d stdlib, %d dep) | "+
-			"batches: %d (avg=%.1f, max=%d)",
-		completed, errs,
-		memHits, memTotal, memHitRate,
-		diskHits, diskTotal, diskHitRate,
-		totalFiles, independent, stdlibOnly, dependent,
-		batchCount, batchAvg, batchMax,
-	)
-}
-
-// isStdlibModule checks if a module path is a standard library module
-func isStdlibModule(path string) bool {
-	switch path {
-	case "option", "result", "bounded", "saturating_cast", "core":
-		return true
-	default:
-		return false
-	}
-}
-
-// classifyFile determines the classification of a file based on its imports
-func classifyFile(builder *ast.Builder, astFile ast.FileID) FileClass {
-	if builder == nil || builder.Files == nil {
-		return FileDependent // Conservative default
-	}
-
-	fileNode := builder.Files.Get(astFile)
-	if fileNode == nil {
-		return FileDependent
-	}
-
-	hasImports := false
-	hasProjectImport := false
-
-	for _, itemID := range fileNode.Items {
-		if imp, ok := builder.Items.Import(itemID); ok {
-			hasImports = true
-			// Get module path from first segment
-			if len(imp.Module) > 0 && builder.StringsInterner != nil {
-				modulePath, _ := builder.StringsInterner.Lookup(imp.Module[0])
-				if !isStdlibModule(modulePath) {
-					hasProjectImport = true
-					break
-				}
-			}
-		}
-	}
-
-	if !hasImports {
-		return FileFullyIndependent
-	}
-	if hasProjectImport {
-		return FileDependent
-	}
-	return FileStdlibOnly
-}
 
 func DiagnoseDirWithOptions(ctx context.Context, dir string, opts DiagnoseOptions, jobs int) (*source.FileSet, []DiagnoseDirResult, error) {
 	files, err := listSGFiles(dir)
@@ -676,193 +541,4 @@ func DiagnoseDirWithOptions(ctx context.Context, dir string, opts DiagnoseOption
 	}
 
 	return fileSet, results, nil
-}
-
-// enrichModuleResults reruns symbol/semantic resolution for directories that
-// declare a multi-file module (pragma module/binary). This ensures all files in
-// the module share one scope when diagnosing a directory.
-func enrichModuleResults(ctx context.Context, _ string, fileSet *source.FileSet, results []DiagnoseDirResult, opts DiagnoseOptions) error {
-	if fileSet == nil {
-		return nil
-	}
-
-	moduleDirs := make(map[string]struct{})
-	pathToIndex := make(map[string]int, len(results))
-
-	for i := range results {
-		res := &results[i]
-
-		path := res.Path
-		if path == "" && res.FileID != 0 {
-			if file := fileSet.Get(res.FileID); file != nil {
-				path = file.Path
-			}
-		}
-		if path == "" {
-			continue
-		}
-
-		pathKey := filepath.ToSlash(path)
-		pathToIndex[pathKey] = i
-		dirKey := filepath.ToSlash(filepath.Dir(pathKey))
-
-		if res.Builder != nil && hasModulePragma(res.Builder, res.ASTFile) {
-			moduleDirs[dirKey] = struct{}{}
-		}
-	}
-
-	if len(moduleDirs) == 0 {
-		return nil
-	}
-
-	typeInterner := types.NewInterner()
-	moduleExports := make(map[string]*symbols.ModuleExports)
-
-	dirKeys := make([]string, 0, len(moduleDirs))
-	for dirKey := range moduleDirs {
-		dirKeys = append(dirKeys, dirKey)
-	}
-	sort.Strings(dirKeys)
-
-	for _, dirKey := range dirKeys {
-		moduleRes, err := diagnoseModuleDirectory(ctx, dirKey, fileSet, opts, typeInterner, moduleExports)
-		if err != nil {
-			return err
-		}
-		for pathKey, res := range moduleRes {
-			idx, ok := pathToIndex[pathKey]
-			if !ok {
-				continue
-			}
-			results[idx].Bag = res.Bag
-			results[idx].FileID = res.FileID
-			results[idx].ASTFile = res.ASTFile
-			results[idx].Builder = res.Builder
-			results[idx].Symbols = res.Symbols
-			results[idx].Sema = res.Sema
-		}
-	}
-
-	return nil
-}
-
-type moduleFileResult struct {
-	Bag     *diag.Bag
-	FileID  source.FileID
-	ASTFile ast.FileID
-	Symbols *symbols.Result
-	Sema    *sema.Result
-	Builder *ast.Builder
-}
-
-// diagnoseModuleDirectory parses and resolves a single directory that declares
-// a multi-file module. It returns per-file diagnostics and semantic results so
-// that directory-level diagnosis can honor shared module scope.
-func diagnoseModuleDirectory(
-	ctx context.Context,
-	dir string,
-	fileSet *source.FileSet,
-	opts DiagnoseOptions,
-	typeInterner *types.Interner,
-	moduleExports map[string]*symbols.ModuleExports,
-) (map[string]moduleFileResult, error) {
-	if fileSet == nil {
-		return nil, nil
-	}
-
-	bag := diag.NewBag(opts.MaxDiagnostics)
-	builder, fileIDs, files, err := parseModuleDir(ctx, fileSet, dir, bag, source.NewInterner(), nil, nil)
-	if err != nil {
-		if errors.Is(err, errModuleNotFound) {
-			return nil, nil
-		}
-		return nil, err
-	}
-
-	reporter := &diag.BagReporter{Bag: bag}
-	meta, ok := buildModuleMeta(fileSet, builder, fileIDs, fileSet.BaseDir(), reporter)
-	if !ok && len(files) > 0 && files[0] != nil {
-		meta = fallbackModuleMeta(files[0], fileSet.BaseDir())
-	}
-	if meta == nil || !meta.HasModulePragma {
-		return nil, nil
-	}
-
-	rec := &moduleRecord{
-		Meta:    meta,
-		Bag:     bag,
-		Builder: builder,
-		FileIDs: fileIDs,
-		Files:   files,
-	}
-
-	if typeInterner == nil {
-		typeInterner = types.NewInterner()
-	}
-	if moduleExports == nil {
-		moduleExports = make(map[string]*symbols.ModuleExports)
-	}
-	if exp := resolveModuleRecord(ctx, rec, fileSet.BaseDir(), moduleExports, typeInterner, opts); exp != nil {
-		moduleExports[normalizeExportsKey(meta.Path)] = exp
-	}
-
-	diagsByFile := make(map[source.FileID][]*diag.Diagnostic)
-	for _, d := range bag.Items() {
-		diagsByFile[d.Primary.File] = append(diagsByFile[d.Primary.File], d)
-	}
-
-	out := make(map[string]moduleFileResult, len(fileIDs))
-	for i, fileID := range fileIDs {
-		file := files[i]
-		fileBag := diag.NewBag(opts.MaxDiagnostics)
-		if ds := diagsByFile[file.ID]; len(ds) > 0 {
-			for _, d := range ds {
-				fileBag.Add(d)
-			}
-		}
-
-		var symRes *symbols.Result
-		if rec.Symbols != nil {
-			if sym, ok := rec.Symbols[fileID]; ok {
-				symCopy := sym
-				symRes = &symCopy
-			}
-		}
-		var semRes *sema.Result
-		if rec.Sema != nil {
-			semRes = rec.Sema[fileID]
-		}
-
-		out[filepath.ToSlash(file.Path)] = moduleFileResult{
-			Bag:     fileBag,
-			FileID:  file.ID,
-			ASTFile: fileID,
-			Symbols: symRes,
-			Sema:    semRes,
-			Builder: builder,
-		}
-	}
-
-	return out, nil
-}
-
-func hasModulePragma(builder *ast.Builder, fileID ast.FileID) bool {
-	if builder == nil || fileID == ast.NoFileID {
-		return false
-	}
-	node := builder.Files.Get(fileID)
-	if node == nil || len(node.Pragma.Entries) == 0 {
-		return false
-	}
-	interner := builder.StringsInterner
-	for _, entry := range node.Pragma.Entries {
-		name, ok := interner.Lookup(entry.Name)
-		if !ok {
-			continue
-		}
-		if name == "module" || name == "binary" {
-			return true
-		}
-	}
-	return false
 }
